Documentation links and refactor module names without underscores
