+++ docs/conf.py
@@ -16,15 +16,15 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath('../python/alf'))
-sys.path.insert(0, os.path.abspath('../python/open_neuro_data'))
-sys.path.insert(0, os.path.abspath('../python/ibllib'))
+sys.path.insert(0, os.path.abspath('../python/openneurodata'))
+sys.path.insert(0, os.path.abspath('../python'))
 print('Python %s on %s' % (sys.version, sys.platform))
 
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'ONE IBL'
+project = 'IBL Library'
 copyright = '2018, International Brain Laboratory'
 author = 'International Brain Laboratory'
 
@@ -43,8 +43,7 @@ release = ''
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc'
-              ]
+extensions = ['sphinx.ext.autodoc']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
