+++ docs/source/ibllib.rst
@@ -1,0 +1,20 @@
+ibllib package
+================
+
+Module contents
+---------------
+
+.. automodule:: ibllib
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+ibllib.webclient module
+-----------------------
+
+.. automodule:: ibllib.webclient
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
