--- docs/source/one_ibl.rst
@@ -1,30 +1,0 @@
-one\_ibl package
-================
-
-Module contents
----------------
-
-.. automodule:: one_ibl
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-one\_ibl.one module
--------------------
-
-.. automodule:: one_ibl.one
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-one\_ibl.utils module
----------------------
-
-.. automodule:: one_ibl.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
