+++ docs/source/oneibl.rst
@@ -1,0 +1,11 @@
+oneibl package
+==============
+
+oneibl.one module
+-----------------
+
+.. automodule:: oneibl.one
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
