+++ python/openneurodata/oneibl/tests/test_one_tutorial.py
@@ -1,0 +1,3 @@
+import oneibl.tutorial_script
+
+print(oneibl.tutorial_script)
